add cloudfront signed url with private key. change swr to check for signed url expiry. split capture repair components to separate files.
