+++ package.json
@@ -18,6 +18,7 @@
     "@auth/prisma-adapter": "^2.9.1",
     "@aws-sdk/client-lambda": "^3.817.0",
     "@aws-sdk/client-s3": "^3.750.0",
+    "@aws-sdk/cloudfront-signer": "^3.821.0",
     "@aws-sdk/s3-request-presigner": "^3.750.0",
     "@dnd-kit/accessibility": "^3.1.1",
     "@dnd-kit/core": "^6.3.1",
