+++ src/app/(signed-in)/capture/[captureId]/start/page.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useReducer } from "react";
 import { redirect, useParams } from "next/navigation";
 import Link from "next/link";
-import useSWR from "swr";
+import useSWR, { mutate } from "swr";
 import { ArrowRight, CircleCheck, FileVideo, Loader2, X } from "lucide-react";
 import { QRCodeSVG } from "qrcode.react";
 
@@ -17,11 +17,17 @@ import {
 import { Button } from "@/components/ui/button";
 
 import { useCapture } from "@/lib/hooks";
-import { CaptureSWROperations, fileFetcher, handleDeleteFile } from "./util";
+import {
+  CaptureSWROperations,
+  fileFetcher,
+  getSWRConfig,
+  handleDeleteFile,
+} from "./util";
 import DeleteUploadDialog from "./components/delete-upload-dialog";
-import { processCaptureFiles, updateCapture } from "@/lib/actions";
+import { ListedFiles, processCaptureFiles, updateCapture } from "@/lib/actions";
 import { cn, Platform } from "@/lib/utils";
 import { CaptureStatus } from "@prisma/client";
+import { isCloudfrontUrlExpired } from "@/lib/aws/s3/client";
 interface CaptureState {
   hasUploads: boolean;
   processingState: "idle" | "pending" | "finished" | "error";
@@ -34,9 +40,9 @@ function captureStateReducer(
   state: CaptureState,
   action:
     | {
-      type: "UPDATE_PROCESS";
-      nextProcessingState: "idle" | "pending" | "finished" | "error";
-    }
+        type: "UPDATE_PROCESS";
+        nextProcessingState: "idle" | "pending" | "finished" | "error";
+      }
     | { type: "UPDATE"; uploadList: any[]; processList: any[] }
 ): CaptureState {
   switch (action.type) {
@@ -65,21 +71,20 @@ function captureStateReducer(
       if (hasUploads) {
         if (state.processingState !== "pending") {
           // determine if all video files have been transcoded
-          const isTranscodeDisabled = (
+          const isTranscodeDisabled =
             !process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA ||
-            process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA === ""
-          );
-          hasTranscoded = isTranscodeDisabled || (
-            hasUploads &&
-            uploadList.every((upload) => {
-              const segments = upload.fileKey.split("/");
-              const fileName = segments.pop() || "";
-              const baseName = fileName.replace(/\.(mp4|mov)$/, "");
-              return processList.some((t) =>
-                t.fileKey.includes(`${baseName}.webm`)
-              );
-            })
-          );
+            process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA === "";
+          hasTranscoded =
+            isTranscodeDisabled ||
+            (hasUploads &&
+              uploadList.every((upload) => {
+                const segments = upload.fileKey.split("/");
+                const fileName = segments.pop() || "";
+                const baseName = fileName.replace(/\.(mp4|mov)$/, "");
+                return processList.some((t) =>
+                  t.fileKey.includes(`${baseName}.webm`)
+                );
+              }));
 
           // determine if all non-video files or thumbnails have been copied
           const nonVideoUploads = uploadList.filter(
@@ -130,17 +135,19 @@ export default function Page() {
   const { data: uploadList = [], isLoading: isUploadListLoading } = useSWR(
     [CaptureSWROperations.UPLOAD_LIST, `uploads/${captureId}`],
     fileFetcher,
-    { refreshInterval: 5000 }
+    getSWRConfig(CaptureSWROperations.UPLOAD_LIST, `uploads/${captureId}`)
   );
 
   const { data: processList = [] } = useSWR(
     [CaptureSWROperations.TRANSCODE_LIST, `processed/${captureId}`],
     fileFetcher,
-    { refreshInterval: 5000 }
+    getSWRConfig(CaptureSWROperations.TRANSCODE_LIST, `processed/${captureId}`)
   );
 
   const handleProcessFiles = async () => {
-    if (uploadList.length === 0) { return; }
+    if (uploadList.length === 0) {
+      return;
+    }
     dispatch({ type: "UPDATE_PROCESS", nextProcessingState: "pending" });
     const res = await processCaptureFiles(uploadList[0].fileKey);
     if (!res.ok) {
@@ -160,7 +167,7 @@ export default function Page() {
     (os === Platform.ANDROID && captureState.hasUploads);
 
   const redirectToTraceProcess = async () => {
-    const captureRes = await updateCapture(captureId,  {
+    const captureRes = await updateCapture(captureId, {
       status: CaptureStatus.PROCESSING,
     });
     if (!captureRes.ok || !captureRes.data || !captureRes.data.id) {
@@ -182,11 +189,9 @@ export default function Page() {
           <CardHeader>
             <CardTitle className="text-2xl">Start capture session</CardTitle>
             <CardDescription>
-              {os == 'android' ? 
-                "Open the ODIM app on your device and scan the QR code below to start the capture session." : 
-                "Open your Camera app and scan the QR code to navigate to the capture session page."
-              }
-
+              {os == "android"
+                ? "Open the ODIM app on your device and scan the QR code below to start the capture session."
+                : "Open your Camera app and scan the QR code to navigate to the capture session page."}
             </CardDescription>
           </CardHeader>
           <CardContent className="p-6 pt-0">
@@ -213,7 +218,11 @@ export default function Page() {
             <CardTitle className="text-2xl">Task recording</CardTitle>
             <CardDescription>
               Your recorded task video will appear here once you start uploading
-              them from your device. <strong>Remember to turn on &ldquo;Do not Disturb&rdquo; before you start recording.</strong>
+              them from your device.{" "}
+              <strong>
+                Remember to turn on &ldquo;Do not Disturb&rdquo; before you
+                start recording.
+              </strong>
             </CardDescription>
           </CardHeader>
           <CardContent>
@@ -227,7 +236,8 @@ export default function Page() {
                       Loading...
                     </>
                   ) : (
-                    `${uploadList.length} file${uploadList.length !== 1 ? "s" : ""
+                    `${uploadList.length} file${
+                      uploadList.length !== 1 ? "s" : ""
                     } uploaded`
                   )}
                 </span>
@@ -250,10 +260,17 @@ export default function Page() {
                         </Link>
                       </div>
                       <DeleteUploadDialog
-                        onContinue={() =>
-                          // TODO: remove from processList too
-                          handleDeleteFile(captureId, file.fileKey)
-                        }
+                        onContinue={() => {
+                          const processedFile = processList.find((p) =>
+                            p.fileKey.includes(file.fileName)
+                          );
+                          console.log(processedFile);
+                          handleDeleteFile(captureId, file.fileKey);
+                          if (processedFile) {
+                            handleDeleteFile(captureId, processedFile.fileKey);
+                          }
+                          dispatch({ type: "UPDATE", uploadList, processList });
+                        }}
                       >
                         <button className="inline-flex items-center cursor-pointer">
                           <X className="size-4 text-neutral-500 hover:opacity-75" />
@@ -274,9 +291,12 @@ export default function Page() {
             {os === Platform.IOS && (
               <div className="flex justify-end mt-4">
                 <Button
-                  disabled={uploadList.length === 0 
-                    || captureState.processingState === "pending"
-                    || captureState.processingState === "finished"}
+                  disabled={
+                    uploadList.length === 0 ||
+                    processList.length > 0 ||
+                    captureState.processingState === "pending" ||
+                    captureState.processingState === "finished"
+                  }
                   onClick={handleProcessFiles}
                 >
                   Submit files
@@ -342,44 +362,45 @@ export default function Page() {
                       <Loader2 className="size-5 text-blue-500 animate-spin" />
                     ))}
                 </li>
-                {(process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA && 
-                  process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA !== "") &&
-                 <li className="flex justify-between items-center px-4 py-2 border-b border-muted-background last:border-none">
-                  <span className="text-sm font-medium">
-                    {captureState.processingState === "idle" &&
-                      (captureState.hasTranscoded
-                        ? "Transcoding complete"
-                        : "Transcode video files")}
+                {process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA &&
+                  process.env.NEXT_PUBLIC_TRANSCODE_LAMBDA !== "" && (
+                    <li className="flex justify-between items-center px-4 py-2 border-b border-muted-background last:border-none">
+                      <span className="text-sm font-medium">
+                        {captureState.processingState === "idle" &&
+                          (captureState.hasTranscoded
+                            ? "Transcoding complete"
+                            : "Transcode video files")}
 
-                    {captureState.processingState === "pending" &&
-                      (captureState.hasTranscoded
-                        ? "Video transcoding complete"
-                        : "Transcoding video files...")}
+                        {captureState.processingState === "pending" &&
+                          (captureState.hasTranscoded
+                            ? "Video transcoding complete"
+                            : "Transcoding video files...")}
 
-                    {captureState.processingState === "finished" &&
-                      (captureState.hasTranscoded
-                        ? "Video transcoding complete"
-                        : "Transcoding video files...")}
-                  </span>
-                  {captureState.processingState === "idle" &&
-                    captureState.hasTranscoded && (
-                      <CircleCheck className="size-5 text-blue-500" />
-                    )}
+                        {captureState.processingState === "finished" &&
+                          (captureState.hasTranscoded
+                            ? "Video transcoding complete"
+                            : "Transcoding video files...")}
+                      </span>
+                      {captureState.processingState === "idle" &&
+                        captureState.hasTranscoded && (
+                          <CircleCheck className="size-5 text-blue-500" />
+                        )}
 
-                  {captureState.processingState === "pending" &&
-                    (captureState.hasTranscoded ? (
-                      <CircleCheck className="size-5 text-blue-500" />
-                    ) : (
-                      <Loader2 className="size-5 text-blue-500 animate-spin" />
-                    ))}
+                      {captureState.processingState === "pending" &&
+                        (captureState.hasTranscoded ? (
+                          <CircleCheck className="size-5 text-blue-500" />
+                        ) : (
+                          <Loader2 className="size-5 text-blue-500 animate-spin" />
+                        ))}
 
-                  {captureState.processingState === "finished" &&
-                    (captureState.hasTranscoded ? (
-                      <CircleCheck className="size-5 text-blue-500" />
-                    ) : (
-                      <Loader2 className="size-5 text-blue-500 animate-spin" />
-                    ))}
-                </li>}
+                      {captureState.processingState === "finished" &&
+                        (captureState.hasTranscoded ? (
+                          <CircleCheck className="size-5 text-blue-500" />
+                        ) : (
+                          <Loader2 className="size-5 text-blue-500 animate-spin" />
+                        ))}
+                    </li>
+                  )}
               </ul>
             </CardContent>
           </Card>
